data-collection/etl/requirements.txt
python-dotenv
pandas
numpy
sqlalchemy
psycopg2-binary
minio
tqdm
